Simosn tester

+++ src/main/java/com/example/kjemsqrecyclebackend/repository/PickupRequestRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PickupRequestRepository extends JpaRepository<PickupRequest, Integer> {
+public interface PickupRequestRepository extends JpaRepository<PickupRequest, Long> {
 }
